Make all exceptions RuntimeExceptions, make Game generic
This change makes all exceptions RuntimeExceptions since none of them
should be checked all of the time (ex. a NoHandsException might be good to
catch in some circumstances since you could skip a player as a result).

This change also makes Game generic. Since Game operates on a type of
dealer and a type of player, the type of dealer and player objects don't need
to be cast.